use purpose build model instead of generic
