+++ apps/llm_providers/apps.py
@@ -1,0 +1,7 @@
+from django.apps import AppConfig
+
+
+class LlmProvidersConfig(AppConfig):
+    name = "apps.llm_providers"
+    label = "llm_providers"
+    default_auto_field = "django.db.models.BigAutoField"
