--- apps/services/apps.py
@@ -1,7 +1,0 @@
-from django.apps import AppConfig
-
-
-class ServicesConfig(AppConfig):
-    name = "apps.services"
-    label = "services"
-    default_auto_field = "django.db.models.BigAutoField"
