+++ gpt_playground/urls.py
@@ -35,7 +35,7 @@ team_urlpatterns = [
     path("", include(web_team_urls)),
     path("team/", include(single_team_urls)),
     path("experiments/", include("apps.experiments.urls")),
-    path("services/", include("apps.services.urls")),
+    path("llm_providers/", include("apps.llm_providers.urls")),
 ]
 
 urlpatterns = [
