+++ templates/teams/components/llm_providers.html
@@ -2,8 +2,8 @@
 {% load form_tags %}
 <div>
   {% if request.team_membership.role == 'admin' %}
-    {% url "services:table" request.team.slug "llm_provider" as table_url %}
-    {% url "services:new_llm" request.team.slug as new_object_url %}
+    {% url "llm_providers:table" request.team.slug as table_url %}
+    {% url "llm_providers:new" request.team.slug as new_object_url %}
     {% include "generic/object_home_content.html" with new_object_url=new_object_url table_url=table_url allow_new=True title="LLM Providers" button_style="btn-outline" %}
   {% else %}
     {% include "generic/object_home_content.html" with new_object_url=new_object_url table_url=table_url allow_new=False title="LLM Providers" button_style="btn-outline" %}
